CMakeLists: don't assume !ARM64 is ARM
CMakeLists.txt currently assume everything not ARM64 to be ARM. When
compiling on x86_64-linux, this pulls in some binaries that are arm32
only.

Some tools make some sense to have accessible even on non-ARM
workstations.  For example, `dtmerge` can be used to merge .dtb files
with overlays on a x86_64 machine (while mounting an sdcard for
example).

Simply by adding this conditionals, we're able to get a `dtmerge` binary
that runs on x86_64-linux.

Signed-off-by: Florian Klink <flokli@flokli.de>